+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added in forge-py fargate terraform code
+- Update and cleaned up github action
 ### Deprecated
 ### Removed
 ### Fixed